refactor: come back top a monolithic file structure
